fix: update task with new api version
I removed the old control commands, which are no longer valid according
to the API. The feedback no longer requires a GET request to be sent
beforehand, because of that there may be some missing fields in the
message we receive. To account for this, I throw an exception whenever
a field is missing, right now I am just ignoring this exception and
continuing with the logic.